+++ io/dbase/dbase.go
@@ -13,15 +13,15 @@ import (
 )
 
 // NewDB creates a new instance of sql.DB using configuration data.
-func NewDB(cnf config.Config) *sql.DB {
-	db, err := sql.Open("postgres", dbUrl(cnf))
+func NewDB(cfg config.Config) *sql.DB {
+	db, err := sql.Open("postgres", dbUrl(cfg))
 	if err != nil {
 		log.Fatalf("Cannot create PostgreSQL connection: %s.", err)
 	}
 	return db
 }
 
-func dbUrl(cnf config.Config) string {
+func dbUrl(cfg config.Config) string {
 	return fmt.Sprintf("postgres://%s:%s@%s:%d/%s?sslmode=disable",
-		cnf.PgUser, cnf.PgPass, cnf.PgHost, cnf.PgPort, cnf.PgDB)
+		cfg.PgUser, cfg.PgPass, cfg.PgHost, cfg.PgPort, cfg.PgDB)
 }